[issue-107] sperated old clip code
Some of the old clip is for an algo we are no longer using. We
want to move this over into geo-spatial eventually, before deleting
it.
